server/pyproject.toml
[project]
name = "tmom"
version = "0.1.0"
description = "Tracker Mom server"
authors = [
    {name = "Paul Bailey", email = "paul@neutron.studio"},
]
dependencies = [
    "django>=5.1.2",
    "django-allauth>=65.1.0",
    "django-allauth-ui>=1.5.0",
    "django-simple-history>=3.7.0",
    "django-ninja>=1.3.0",
    "wonderwords>=2.2.0",
    "pyjwt>=2.9.0",
]
requires-python = "==3.11.*"
readme = "README.md"
license = {text = "AGPL-3.0"}


[tool.pdm]
distribution = false

[tool.pdm.scripts]
dev = "python manage.py runserver 0.0.0.0:8000"
test = "pytest -sx"

[tool.ruff]
line-length = 100
indent-width = 2

[dependency-groups]
dev = ["pytest-django>=4.9.0"]

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "tmom.settings.test"
python_files = ["tests.py", "test_*.py", "*_test.py", "testing/python/*.py"]
